Move wcs to it's own file
